✨ Boutgifter: Ta hensyn til makssats i beregningen
