+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/boutgifter/beregning/BoutgifterBeregnBeløpUtil.kt
@@ -1,16 +1,13 @@
 package no.nav.tilleggsstonader.sak.vedtak.boutgifter.beregning
 
-object BoutgifterBeregnBeløpUtil {
-//    private val PROSENT_50 = BigDecimal(0.5)
-//    private val PROSENTGRENSE_HALV_SATS = 50
+import kotlin.math.min
 
-    fun beregnBeløp(utgifter: Map<Unit, List<UtgiftBeregning>>): Int {
-//        if (studieprosent <= PROSENTGRENSE_HALV_SATS) {
-//            return BigDecimal(sats)
-//                .multiply(PROSENT_50)
-//                .setScale(0, RoundingMode.HALF_UP)
-//                .toInt()
-//        }
-        return TODO("Beregning av beløp med utgifter og makssats for boutgifter er ikke implementert enda")
+object BoutgifterBeregnBeløpUtil {
+    fun beregnBeløp(
+        utgifter: Map<Unit, List<UtgiftBeregning>>,
+        makssats: Int,
+    ): Int {
+        val summerteUtgifter = utgifter.values.flatten().sumOf { it.utgift }
+        return min(summerteUtgifter, makssats)
     }
 }
